Stabilize GC Tests
The callback g_object_weak_ref could be GCed while it can be still called
from the native side. This leads to a segfault.

The implementation makes the native callback a static member of GCTracker
and thus ensures that is not collected. The back mapping to the correct
GCTracker is done via identityHashCode.