Don't introduce additional reference for floating reference
gst_element_factory_make and gst_element_factory_create create elements
with floating references. Native#objectFor sinks the reference and so
no additional ref is needed.

gst_bin_new creates the object with a floating reference. This is
correctly handled by the initializer and does not need an additional
reference.