Set minimal source compatibility
This is simple change to make project green in Eclipse after
import is done. In Eclipse immediately after POM is imported
the project turns red because Eclipse assumes 1.5 source
compatibility (not sure why). This change is to fix this.
With this change, immediately after POM is imported, the
source compatibility will be set to 1.6 which is a minimal
Java version for this project.